cjson tests to a separate class
